refactor: update the Toaster component
