--- frontend/src/components/PageToast/types.ts
@@ -1,6 +1,0 @@
-import { ExternalToast } from "sonner";
-
-export interface IPageToastProps {
-    message: string;
-    options?: Omit<ExternalToast, "description">;
-}
